Dereferenceable is now in the Concern module.
